+++ multi-gammon-core/src/main/java/org/akoshterek/backgammon/agent/ETraceEntry.java
@@ -10,15 +10,15 @@ import java.util.Arrays;
  *         date: 05.12.2015
  */
 public class ETraceEntry {
-    public ETraceEntry(double[] input, PositionClass pc, AuchKey auch) {
+    public ETraceEntry(float[] input, PositionClass pc, AuchKey auch) {
         this.input = Arrays.copyOf(input, input.length);
         this.pc = pc;
         this.auch = new AuchKey(auch);
-        eTrace = 1.0;
+        eTrace = 1.0f;
     }
 
-    public double[] input;
+    public float[] input;
     public PositionClass pc;
     public AuchKey auch;
-    public double eTrace;
+    public float eTrace;
 }
